requirements.txt
streamlit
pandas
numpy
sentence-transformers
pymupdf
pdfminer.six
rapidfuzz
torch

